Fix some compiler warnings, double->float in a few place.  Prevent crash with effect of lenght 1 frame and color gradient set.

+++ xLights/RenderBuffer.h
@@ -334,18 +334,18 @@ public:
     PaletteClass& GetPalette() { return palette; }
 
     HSVValue Get2ColorAdditive(HSVValue& hsv1, HSVValue& hsv2);
-    double GetEffectTimeIntervalPosition();
-    double GetEffectTimeIntervalPosition(float cycles);
+    float GetEffectTimeIntervalPosition();
+    float GetEffectTimeIntervalPosition(float cycles);
 
 
 
     void CopyPixelsToDisplayListX(Effect *eff, int y, int sx, int ex, int inc = 1);
     // must hold the lock and be sized appropriately
-    void SetDisplayListHRect(Effect *eff, int startIdx, double x1, double y1, double x2, double y2,
+    void SetDisplayListHRect(Effect *eff, int startIdx, float x1, float y1, float x2, float y2,
                              const xlColor &cx1, const xlColor &cx2);
-    void SetDisplayListVRect(Effect *eff, int startIdx, double x1, double y1, double x2, double y2,
+    void SetDisplayListVRect(Effect *eff, int startIdx, float x1, float y1, float x2, float y2,
                              const xlColor &cy1, const xlColor &cy2);
-    void SetDisplayListRect(Effect *eff, int startIdx, double x1, double y1, double x2, double y2,
+    void SetDisplayListRect(Effect *eff, int startIdx, float x1, float y1, float x2, float y2,
                             const xlColor &cx1y1, const xlColor &cx1y2,
                             const xlColor &cx2y1, const xlColor &cx2y2);
